+++ src/App.js
@@ -3,7 +3,7 @@ import { Navigation } from "./common/Navigation";
 import { PageList } from "./common/PageList";
 import { MoviesList } from "./features/MoviesList";
 import { PeopleList } from "./features/PeopleList";
-import { NoResult } from "./features/NoResults";
+import { NoResult } from "./common/NoResults";
 
 export const App = () => (
   <>
